pyproject.toml
[project]
name = "instacart-online-grocery-basket-analysis-data-engineering-project"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "google-cloud-bigquery>=3.32.0",
    "google-cloud-storage>=2.18.0",
    "kaggle>=2.0.0",
    "kagglehub>=1.0.0",
    "pandas>=3.0.1",
    "pyarrow>=23.0.1",
]
